refactor(commands): move the stats logic to a service
Instead of the function getServerMemberCount() being on the commands.ts file, we move that logic to it's own file and create an interface.  Then an object of that type is injected in the commands function (dependency injection).

With this change, the commands module is abstracted of the way that data is fetched. This is important since we want to save that data to a DB, and read from it. For that, we only need to add a class that implements the StatsService interface.

Issue: #37