feat added (roles -a)
Introduced a more fragmented structure, where there are multiple ways of accessing the role-list. The default on (roles) is going to be a list of Assignable roles, the (roles -a) or (roles all) would give a list of all roles. (roles <rolename>) would give the description for the role, if it exists.