feat(roles): inline and improve formating (#432)
* feat added (roles -a)

Introduced a more fragmented structure, where there are multiple ways of accessing the role-list. The default on (roles) is going to be a list of Assignable roles, the (roles -a) or (roles all) would give a list of all roles. (roles <rolename>) would give the description for the role, if it exists.

* fix: perfected `roles -a` to give conditional roles at the bottom

* feat: roles extension

The roles command can now be used with arguments, such as a role-name or all or -a. The command by default now returns a list of assignable roles, rather than all roles

* fix: corrected type issues and failing build

* fix: attempt to fix CI

I hope this fixes the build errors. It seems like the error is only in the build, but locally can't
be reproduced. I have converted the list to a string, rather than how it was previously directly
mentioned.

fix #3455

* fix: ci fix attempt

* refactor(roles.ts): reduced no. of iterations in `roleList`

As pointed out by @nhcarrigan in the discord chat(#help), the rolesList iterates over the list 4
times, which is (yikes).. reduced one unneccessary .filter()

* refactor(roles.ts): implemented suggestion to reduce iterative complexity

* refactor(roles.ts): more readable single line code

Took into account the suggestions from @eddiejaoude

* refactor(roles.ts): changed phrasing on example uasge

type- *usage. Takes into account suggestion from BOLT04

* refactor(roles.ts): implemented function, cleared mess

This commit takes into account the suggestions given by @BOLT04(David). This makes the code more
readable and reduces redundancy in the code.

* suggestion(BOLT04): destructuring value from role-list function

Co-authored-by: David <didicp4@gmail.com>

* fix(roles.ts): imported types from discordjs

types Role and Collection were missing.

* refactor(roles.ts): implemented destructuring

* refactor(roles.ts): changed naming of isValidRole param

It's a bit weird that we return if the role is valid (it should be the opposite. `isValidRole` has
been replaced with `notValid`

345

* refactor(roles.ts): uniform naming for arg[1]

Co-authored-by: David <didicp4@gmail.com>
Co-authored-by: Allan Regush <17693494+AllanRegush@users.noreply.github.com>